Removed redundant super-interfaces. (cherry picking from commit f2bda96580304280f7d367a3823c6ded9b6a7b2c) Removed unecessary throw declarations and catch blocks. (cherry picking from commit e9cf0f61e7128e4bbc73ae2218b7ae34ecb4c846) Replaced indirect access to static members by direct access. (cherry picking from commit 604ba4a96c7afbc51aa1735dff1a8d6befa014f1) Qualified access to instance members. (cherry picking from commit c2e7e264b1e2271791c73bbfe7ce4fb6347a5e1f) Removed unecessary casts. (cherry picking from commit fcb9103f539416dadfefd10fe5cad1da73642d77) Removed unused imports. (cherry picking from commit 00fa0639702e7df84e040e2e06fea31e7759a157) Removed redundant null checks. (cherry picking from commit 129a8cda8e19fb6f5e7a457a7da84dda308d959a) Removed unnecessary else blocks. (cherry picking from commit 56955d3c4b7ccf12622f0d2ace68cf3021a144c3) Fixed hidden fields. (cherry picking from commit dead701244b4608cbaa2e6898f8cfc3a06bc9594) Fixed unused local variables and members issues. (cherry picking from commit 14b58af1623f7bb56f3de8826d9069eadc63c91c)

+++ src/helma/util/Base64.java
@@ -60,15 +60,15 @@ static public char[] encode(byte[] data)
         boolean quad = false;
         boolean trip = false;
 
-        int val = (0xFF & (int) data[i]);
+        int val = (0xFF & data[i]);
         val <<= 8;
         if ((i+1) < data.length) {
-            val |= (0xFF & (int) data[i+1]);
+            val |= (0xFF & data[i+1]);
             trip = true;
         }
         val <<= 8;
         if ((i+2) < data.length) {
-            val |= (0xFF & (int) data[i+2]);
+            val |= (0xFF & data[i+2]);
             quad = true;
         }
         out[index+3] = alphabet[(quad? (val & 0x3F): 64)];
